Fold test executor abstraction
